Remove nearly all downsampling code
Downsampling code came in when the code was forked from Thanos, but
Mimir documents that you cannot use this feature.

This PR removes all references to `thanos/pkg/compact/downsample`,
including some tests that are primarily checking behaviour with
downsampled data.

I left some references to the `Downsample` field in block metadata,
mostly checking that it is zero or printing it out for debugging.