Ruler, Querier: Increase frequency of gRPC healthchecks
This uses the same frequency and timeout that the querier has to the
query-frontends. This will help in cases when the store-gateways become
unresponsive for prolonged periods of time. The querier will more
quickly remove the connection to that store-gateways. So fewer new
queries will go to it and existing queries will fail faster.

Does the same for the ruler-ruler clients (used to get state of running
rules in other rulers) for good measure.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>